# src/api/rest/docs/swagger.yaml
basePath: /tumblebug
definitions:
...
  model.CustomImageStatus:
    enum:
    - Available
    - Unavailable
    type: string
    x-enum-varnames:
    - MyImageAvailable
    - MyImageUnavailable
...
  model.IdList:
    properties:
      output:
        items:
          type: string
        type: array
    type: object
...
  model.KeyValue:
    properties:
      key:
        type: string
      value:
        type: string
    type: object
...
  model.SimpleMsg:
    properties:
      message:
        example: Any message
        type: string
    type: object
...
  model.TbCustomImageInfo:
    properties:
      associatedObjectList:
        items:
          type: string
        type: array
      connectionName:
        example: aws-ap-southeast-1
        type: string
      creationDate:
        example: "2022-10-18T08:12:48Z"
        type: string
      cspCustomImageId:
        example: ami-06eb41e14121c550a
        type: string
      cspCustomImageName:
        example: aws-ap-southeast-1-snapshot
        type: string
      description:
        type: string
      guestOS:
        description: Windows7, Ubuntu etc.
        type: string
      id:
        example: aws-ap-southeast-1-snapshot
        type: string
      isAutoGenerated:
        type: boolean
      keyValueList:
        items:
          $ref: '#/definitions/model.KeyValue'
        type: array
      name:
        example: aws-ap-southeast-1-snapshot
        type: string
      namespace:
        description: required to save in RDB
        example: default
        type: string
      sourceVmId:
        example: aws-ap-southeast-1-1
        type: string
      status:
        allOf:
        - $ref: '#/definitions/model.CustomImageStatus'
        example: Available
      systemLabel:
        description: SystemLabel is for describing the Resource in a keyword (any
          string can be used) for special System purpose
        example: Managed by CB-Tumblebug
        type: string
      uuid:
        description: uuid is universally unique identifier for the resource
        type: string
    type: object
  model.TbCustomImageReq:
    properties:
      connectionName:
        type: string
      cspCustomImageId:
        description: This field is for 'Register existing custom image'
        type: string
      description:
        type: string
      name:
        type: string
      sourceVmId:
        type: string
    required:
    - name
    type: object
...
  resource.JSONResult:
    type: object
...
  resource.RestGetAllCustomImageResponse:
    properties:
      customImage:
        items:
          $ref: '#/definitions/model.TbCustomImageInfo'
        type: array
    type: object
...
info:
  contact:
    name: API Support
    url: https://github.com/cloud-barista/cb-tumblebug/issues/new/choose
  description: CB-Tumblebug is an open source system for managing multi-cloud infrastructure
    consisting of resources from multiple cloud service providers. (Cloud-Barista)
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://github.com/cloud-barista/cb-tumblebug/blob/main/README.md
  title: CB-Tumblebug REST API
  version: latest
paths:
...
  /ns/{nsId}/resources/customImage:
    delete:
      consumes:
      - application/json
      description: Delete all customImages
      operationId: DelAllCustomImage
      parameters:
      - default: default
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - default: ""
        description: Delete resources containing matched ID-substring only
        in: query
        name: match
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.IdList'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Delete all customImages
      tags:
      - '[Infra Resource] Image Management'
    get:
      consumes:
      - application/json
      description: List all customImages or customImages' ID
      operationId: GetAllCustomImage
      parameters:
      - default: default
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - description: Option
        enum:
        - id
        in: query
        name: option
        type: string
      - description: Field key for filtering (ex:guestOS)
        in: query
        name: filterKey
        type: string
      - description: 'Field value for filtering (ex: Ubuntu18.04)'
        in: query
        name: filterVal
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Different return structures by the given option param
          schema:
            allOf:
            - $ref: '#/definitions/resource.JSONResult'
            - properties:
                '[DEFAULT]':
                  $ref: '#/definitions/resource.RestGetAllCustomImageResponse'
                '[ID]':
                  $ref: '#/definitions/model.IdList'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: List all customImages or customImages' ID
      tags:
      - '[Infra Resource] Image Management'
    post:
      consumes:
      - application/json
      description: Register existing Custom Image in a CSP (option=register)
      operationId: PostCustomImage
      parameters:
      - default: default
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - description: 'Option: '
        enum:
        - register
        in: query
        name: option
        required: true
        type: string
      - description: Request to Register existing Custom Image in a CSP
        in: body
        name: customImageRegisterReq
        required: true
        schema:
          $ref: '#/definitions/model.TbCustomImageReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TbCustomImageInfo'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Register existing Custom Image in a CSP
      tags:
      - '[Infra Resource] Image Management'
  /ns/{nsId}/resources/customImage/{customImageId}:
    delete:
      consumes:
      - application/json
      description: Delete customImage
      operationId: DelCustomImage
      parameters:
      - default: default
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - description: customImage ID
        in: path
        name: customImageId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Delete customImage
      tags:
      - '[Infra Resource] Image Management'
    get:
      consumes:
      - application/json
      description: Get customImage
      operationId: GetCustomImage
      parameters:
      - default: default
        description: Namespace ID
        in: path
        name: nsId
        required: true
        type: string
      - description: customImage ID
        in: path
        name: customImageId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TbCustomImageInfo'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.SimpleMsg'
      summary: Get customImage
      tags:
      - '[Infra Resource] Image Management'
...
securityDefinitions:
  BasicAuth:
    type: basic
  Bearer:
    description: Type "Bearer" followed by a space and JWT token ([TBD] Get token
      in http://xxx.xxx.xxx.xxx:xxx/auth)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"